Smoother Textbox; Fixed Tileboost Saving; UseStyle now resets to default value when Right Clicked; Fixed UseStyle Textbox position when in Limited Mode

+++ ItemModifications.cs
@@ -461,7 +461,7 @@ namespace ItemModifier
             }
             if (item.tileBoost != defaultItem.tileBoost)
             {
-                tag.Add("TileBoost", (byte)item.tileBoost);
+                tag.Add("TileBoost", (byte)(sbyte)item.tileBoost);
             }
             if (item.buffType != defaultItem.buffType)
             {
